fix: pager scroll math — u16 saturation, wrap-row reachability, pane scroll sync
PR4 of CODE_REVIEW_FOLLOWUP — 8 findings across 3 root causes (7 fixed, 1
documented).

u16 saturation (mod.rs:141, pager_handler/mod.rs:311):
- PagerView.scroll (and saved_alt_scroll) widened u16 → usize so pagers over
  65535 lines are scrollable. All scroll-domain math widened with it;
  viewport_height stays u16. Persistence (pager_positions) u16 → u64
  (backward-compatible; old small values still parse).

wrap-row math:
- visual_rows (layout.rs:131) rewritten to mirror wrap_line's greedy fill
  (wide-char waste) instead of div_ceil — scroll_max no longer clamps short
  of the bottom on wide-char content.
- scroll_to_keep_visible (selection.rs:281) is now wrap-aware: counts visual
  rows so the visual cursor can't slide off the bottom under wrap; O(vh).
- wrap_line_capped (render.rs:180) bounds per-frame wrapping to the visible
  viewport budget; wrap_line delegates with usize::MAX (byte-identical).

pane vt100 scroll (pane/mod.rs:472, :518):
- max_scrollback probes the real buffered length (scrollback_len) instead of
  a hardcoded 10_000; apply_scroll reads vt100's clamped offset back into
  scroll_offset, killing the post-'g' "scroll-down dead" dead zone.

Documented limitation (render.rs:125): scroll is logical-line-granular, so a
single logical line taller than the viewport can't be scrolled through under
wrap (needs visual-row-granular scrolling — a scroll-model rearchitecture).
Noted on PagerView::wrap with mitigations.

4 new pager unit tests; pane scroll is live-pty (owner-tested). Gate green.

+++ src/app/mod.rs
@@ -347,14 +347,14 @@ enum PagerReturn {
     TempFile {
         path: PathBuf,
         title: String,
-        scroll: u16,
+        scroll: usize,
         mount: crate::ui::pager::Mount,
         pane_scroll: bool,
     },
     /// On-disk file: reopen from the original path.
     SourceFile {
         path: PathBuf,
-        scroll: u16,
+        scroll: usize,
         mount: crate::ui::pager::Mount,
         pane_scroll: bool,
     },
